Merge release 27 may2022 https://jira.carecloud.com/projects/BREEZ/versions/14654

+++ CarePay/CarePayLibrary/src/main/java/com/carecloud/carepaylibray/appointments/models/ScheduleAppointmentRequestDTO.java
@@ -28,7 +28,7 @@ public class ScheduleAppointmentRequestDTO {
         private String endTime;
 
         @SerializedName("location_id")
-        private int locationId;
+        private double locationId;
 
         private transient String locationGuid;
 
@@ -68,11 +68,11 @@ public class ScheduleAppointmentRequestDTO {
             this.endTime = endTime;
         }
 
-        public int getLocationId() {
+        public double getLocationId() {
             return locationId;
         }
 
-        public void setLocationId(int locationId) {
+        public void setLocationId(double locationId) {
             this.locationId = locationId;
         }
